Add Command-R+ model support (not to be confused with Command-R).
This model is almost the same as Command-R, but has two norm layers in
the attention.

This commit adds detection in convert-hf-to-gguf.py script (the models
use the same Transformers class so it needs hacks to distinguish it from
Command-R).

Added code in llama.cpp that runs the same codepath as Command-R, just
with the layernorms when using the plus model.