Band-aid fix for Command-R+ model to avoid quantization overflow bug.
There is a bug in all quantization functions in the sense that they
accept 'int k' as the argument that tells how many values to quantize.

int k is only 32-bit signed integer. Command-R+ has a tensor that's
larger than can fit inside this 'int k'. The proper solution would
likely be to upgrade the entire quant battery to size_t or int64_t or
similar.

I'm on a goal just to get this one model running so I went with a
band-aid fix here.