dots1 model support (WIP).
This is the state of the code as it was when I managed to get some
coherent text out of it.

This needs cleaning up and verification with HF version. The dots1 model
is a patchwork of qwen3 and deepseekv3 pieces, and this commit is made
out of haphazard copypasting (and no comments where I did that).

This IMO a bit simpler model to make run because it uses ready-made
components from other models as-is. Or well, at least me assuming that
is true was enough to get coherent text out of it.

I ran it on Mac Studio 192GB, Metal. With a short prompt and 1000 max
context length.

I hacked the convert_hf_to_gguf.py to work to make a bf16 quant, and
then made a Q4_K to test with.