Clean dots1 architecture additions in convert_hf_to_gguf.py
I checked which codepaths are actually used in the conversion, and
removed extra parts. The code was more or less copied verbatim from
DeepseekV2 converter and it had parts the Dots cannot handle yet (e.g.
MLA stuff).

Code is now shorter. I also added some comments, anticipating if future
dots models might change some things.

As of making this commit, I have not tested that this cleanup actually
still functions and does not crash or produce bad .ggufs. If you are
someone reading this commit and eagerly making .ggufs and it doesn't
work, try going to earlier commits in this dots line. I plan to make a
separate branch and squash these commits at the end when making a pull
request.